Move functions to parent component, update tests
