+++ web/app/components/header/facet-dropdown-list.hbs
@@ -1,6 +1,6 @@
 {{on-document "keydown" this.onKeydown}}
 <div
-  {{did-insert this.registerPopover}}
+  {{did-insert @registerPopover}}
   role={{if @inputIsShown "combobox"}}
   id="facet-dropdown-popover"
   class="facet-dropdown-popover
@@ -12,22 +12,22 @@
     <div class="relative px-1 pb-1 border-b border-b-color-border-faint">
       <Hds::Form::TextInput::Base
         {{did-insert this.registerInput}}
-        {{on "input" (perform this.onInput)}}
-        {{on "focusin" this.resetMenuItemIndex}}
-        @value={{this.query}}
+        {{on "input" @onInput}}
+        {{on "focusin" @resetMenuItemIndex}}
+        @value={{@query}}
         @type="search"
         aria-controls="facet-dropdown-menu"
         class="leading-none"
         placeholder="Filter..."
         class="facet-dropdown-input"
         aria-activedescendant={{unless
-          (eq this.focusedItemIndex -1)
-          (concat "facet-dropdown-menu-item-" this.focusedItemIndex)
+          (eq @focusedItemIndex -1)
+          (concat "facet-dropdown-menu-item-" @focusedItemIndex)
         }}
       />
     </div>
   {{/if}}
-  <div class="overflow-auto h-full relative" {{did-insert this.registerScrollContainer}}>
+  <div class="overflow-auto h-full relative" {{did-insert @registerScrollContainer}}>
     {{#if this.noMatchesFound}}
       <div
         data-test-facet-dropdown-menu-empty-state
@@ -42,18 +42,18 @@
         role={{if @inputIsShown "listbox" "menu"}}
         class="py-3 list-none"
         aria-activedescendant={{unless
-          (or @inputIsShown (eq this.focusedItemIndex -1))
-          (concat "facet-dropdown-menu-item-" this.focusedItemIndex)
+          (or @inputIsShown (eq @focusedItemIndex -1))
+          (concat "facet-dropdown-menu-item-" @focusedItemIndex)
         }}
       >
-        {{#each-in this.shownFacets as |value attrs|}}
+        {{#each-in @shownFacets as |value attrs|}}
           <Header::FacetDropdownListItem
             @role={{if @inputIsShown "option" "menuitem"}}
             @value={{value}}
             @selected={{attrs.selected}}
             @count={{attrs.count}}
-            @setFocusedItemIndex={{this.setFocusedItemIndex}}
-            @focusedItemIndex={{this.focusedItemIndex}}
+            @setFocusedItemIndex={{@setFocusedItemIndex}}
+            @focusedItemIndex={{@focusedItemIndex}}
           />
         {{/each-in}}
       </ol>
